Make store credits non-expiring by default
This is essentially how store credits behave currently. We have an
"expiring" type, but without an expiration they don't really expire.

If we continue to make expiring the default, we have to worry about what
to set the default expiration date to, since there's now a validation
that requires it for expiring credit types.