Autosave creditable association
In `Spree::ReimbursementType::ReimbursementHelpers` we build the
creditable type first and assign it to the Spree::Reimbursement::Credit.
However, when we save the credit and the creditable is not valid, it
fails silently because it doesn't autosave the association.

This can lead to confusing errors because we expect `create_credit` to
give us back a valid credit, but instead we get a persisted credit with
an invalid creditable association.